starter/requirements.txt
altgraph==0.17.2
Brotli==1.0.9
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
multivolumefile==0.2.3
mysql-connector-python==8.0.28
pefile==2021.9.3
protobuf==3.20.0
py7zr==0.18.3
pybcj==0.5.1
pycryptodomex==3.14.1
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
pyppmd==0.18.2
pywin32-ctypes==0.2.0
pyzstd==0.15.2
smmap==5.0.0
texttable==1.6.4
zipfile-deflate64==0.2.0
